Use vim tabs instead of barbar tabs.
The latter is not very useful for my workflow. vim tabs allow me to have
pages of window layouts. barbar hijacks the tab area to list all
buffers, and decided I do not like this as I can easily select any
buffer with Telescope or similar, but selecting a particular tab/window
layout needs to be possible too.