Remove Paredit
I was using it solely for splice-killing-backward, but it never worked
correctly in comparison with the Emacs paredit, and even sometimes
leaves unbalanced parentheses when used in combination with parinfer
like I am. This function is not used that often by me, so it is worth
removing support for paredit until I feel like reimplementing that
functionality myself.